+++ lib/rspec/core/example.rb
@@ -266,13 +266,8 @@ module RSpec
       end
 
       # @private
-      def any_apply?(filters)
-        MetadataFilter.any_apply?(filters, metadata)
-      end
-
-      # @private
-      def all_apply?(filters)
-        MetadataFilter.all_apply?(filters, metadata) || @example_group_class.all_apply?(filters)
+      def apply?(predicate, filters)
+        MetadataFilter.apply?(predicate, filters, metadata)
       end
 
       # @private