+++ lib/rspec/core/example_group.rb
@@ -536,13 +536,8 @@ module RSpec
       end
 
       # @private
-      def self.any_apply?(filters)
-        MetadataFilter.any_apply?(filters, metadata)
-      end
-
-      # @private
-      def self.all_apply?(filters)
-        MetadataFilter.all_apply?(filters, metadata)
+      def self.apply?(predicate, filters)
+        MetadataFilter.apply?(predicate, filters, metadata)
       end
 
       # @private