+++ lib/rspec/core/filter_manager.rb
@@ -217,7 +217,7 @@ module RSpec
       end
 
       def include_example?(example)
-        @rules.empty? ? true : example.any_apply?(@rules)
+        @rules.empty? ? true : example.apply?(:any?, @rules)
       end
 
       def standalone?
@@ -252,7 +252,7 @@ module RSpec
       }.freeze
 
       def include_example?(example)
-        example.any_apply?(@rules) || example.any_apply?(CONDITIONAL_FILTERS)
+        example.apply?(:any?, @rules) || example.apply?(:any?, CONDITIONAL_FILTERS)
       end
     end
   end